\name{npst-package}
\alias{npst-package}
\docType{package}
\title{
Implementation of a Generalized Nonparametric Seasonality Test  
}
\description{
Package 'npst' implements a generalization of Hewitt's (et al. 1971) and
Rogerson's (1996) test for seasonality.
}
\details{
\tabular{ll}{
Package: \tab npst\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-11-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Roland Rau

Maintainer: Roland Rau <roland.rau@gmail.com>
}
\references{
  D. Hewitt, J. Milner, A. Csima, A. Pakula (1971): On Edwards'
  Criterion of Seasonality and a Non-Parametric Alternative. British
  Journal of Preventive Social Medicine 25:174--176
      
  P.A. Rogerson (1996): A Generalization of Hewitt's Test for
  Seasonality. International Journal of Epidemiology 25:644--648

  S.D. Walter (1980): Exact significance levels for Hewitt's test for
  seasonality. Journal of Epidemiology and Community Health
  34:147--149
}
\keyword{ package }
\examples{
npst()
}
